Moved all exports to Emulator so that helper no more imports Emulator; Fixed displaying asset token name in utxos
